+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * the parameter missing_join supports to NULL in woebin function
 * modified the oot param names in split_df function
+* the woebin_adj function supports more params for binning
 
 # scorecard 0.4.3